requirements.txt
google-generativeai>=0.5.0
PyPDF2>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
gunicorn>=21.0.0
openai>=1.12.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
